treat "SSSSSS" in SimpleDateFormat as microseconds
This patch allows users to choose timestamp precision in a simple
manner.  The SimpleDateFormat pattern "SSS" will be replaced with the
fractional second part of the entry's timestamp with millisecond
precision (as before), and "SSSSSS" with microsecond precision.

Since runs of "S" characters other than 3 or 6 would not render
correctly (without this patch, "s.SSSS" will render 3 seconds, 123
milliseconds as "3.0123"), such patterns are passed through verbatim so
the user can realize there is a problem.

Signed-off-by: Nathaniel W. Turner <nate@houseofnate.net>