LOGCXX-394: With removing static in Level::get* some of the tests failed because they relied on the fact that LevelPtr overrides operator== by comparing pointers and because of the static levels those pointers where always the same for every level. I changed this by LevelPtr overriding operator== on it's own and comparing the level values instead of it's pointers. The seems to make more sense for Levels anyways, especially because Level::equals already behaved that way and ignored the pointer of LevelPtr, but compared it's level value instead.
I'll focus on == and != for now and not override each method of ObjectPtr because those 2 are only used in Level.

We don't have many options anyways, either we revert to using sttaic levels and introduce the bug again or we need to adopt already written code to compare using a cast or such.