LOGCXX-457: I think I've found the reason for at least test6 failing: The first rollover is done on the second log statement, which is only issued after some sleeping time. The problem is that the first file name to check for is created using "now" and that could have easily past if the second statement is issued. If the first file name is created with that in mind one second in the future, the first log statement can consume the current second of "now" including the sleep and still the first file will be found successfully, because it is expected to be created with the next second.
I've added a waitFactor to logMsgAndSleep, because I think at least for test6 we need to make sure that each subsequent file is created in the next second, because that's how we build the file names: One second each. If we sleep in the logging only for half a second, two we may have two log statements within the same second resulting in another missing file. Interestingly, during my tests that didn't seem to make any difference...

Additionally I've created test7 repeatly calling test6, because during my tests test6 succeeded even without my changes sometimes, but not always. With the changes it seems to succeed always.