LOGCXX-523 Call error handler if rollover fails (#64)
* LOGCXX-523 Call error handler if rollover fails

If the rollover fails, call the specified error handler for
the given appender.  This means that you can now replace appenders
with a new appender.

The entire API at this point is not very robust at the moment.  A
better error handling mechanism needs to be made to properly
handle the errors so that the error handler can have enough
data to do its job.

* Fix build on Windows

* Add include file needed for windows