LOGCXX-431 (#68)
* LOGCXX-431 Update thread creation

In order to make it possible to block signals when a thread is
created, a new API to let users configure what to do when a
thread is started has been created.

Documentation on how Log4cxx handles threads, and information
about good ways to handle signals has been added as well.

Note that this does not automatically configure signals to be blocked
by default on POSIX systems.  This is intended to be the default
shortly.