Fix constructions of sed filter in corner cases
Some tests use sed to generate predictable outputs and employ sed files for
that.  In some circumstances, the generated sed file syntax is incorrect, if
the subsitions data contains the sed expresssion seperator. The patch escapes
the data if needed.  The bug manifested in (local) autopkgtest runs, when the
temporary directory contained the separator ("Q").