Fix constructions of SED filter in corner cases (#83)
* Fix constructions of SED filter in corner cases

Some tests use SED to generate predictable outputs and employ SED files for
that.  In some circumstances, the generated SED file syntax is incorrect, if
the substitution data contains the SED expression separator. The patch escapes
the data if needed. The bug manifested in (local) autopkgtest runs, when the
temporary directory contained the separator ("Q").

Co-authored-by: Tobias Frost <tobi@debian.org>